[Refactor] Refactor styled-components to style files
